+++ lib/documents/models/nodes/container.model.dart
@@ -4,10 +4,10 @@ import '../style.model.dart';
 import 'child-query.model.dart';
 import 'node.model.dart';
 
-// Container can accommodate other nodes.
-// Delegates insert, retain and delete operations to children nodes. For each
-// operation container looks for a child at specified index position and
-// forwards operation to that child.
+// Container can accommodate other document nodes.
+// Delegates insert, retain and delete operations to children nodes.
+// For each operation container looks for a child at specified
+// index position and forwards operation to that child.
 // Most of the operation handling logic is implemented by [Line] and [Text].
 abstract class ContainerM<T extends NodeM?> extends NodeM {
   final LinkedList<NodeM> _children = LinkedList<NodeM>();