+++ lib/documents/services/document.service.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
@@ -5,6 +7,7 @@ import '../../blocks/services/lines-blocks.service.dart';
 import '../../blocks/widgets/editable-text-block.dart';
 import '../../shared/state/editor.state.dart';
 import '../models/attribute.model.dart';
+import '../models/document.model.dart';
 import '../models/nodes/block.model.dart';
 import '../models/nodes/line.model.dart';
 import 'delta.utils.dart';
@@ -19,15 +22,19 @@ class DocumentService {
 
   DocumentService._privateConstructor();
 
-  List<Widget> docBlocsAndLines({required EditorState state}) {
-    final docElems = <Widget>[];
+  // Renders all the text elements (lines and blocs) that are visible in the editor.
+  List<Widget> documentBlocsAndLines({
+    required DocumentM document,
+    required EditorState state,
+  }) {
+    final docElements = <Widget>[];
     final indentLevelCounts = <int, int>{};
-    final nodes = state.document.document.root.children;
+    final nodes = document.root.children;
 
     for (final node in nodes) {
       // Line
       if (node is LineM) {
-        docElems.add(
+        docElements.add(
           Directionality(
             textDirection: getDirectionOfNode(node),
             child: _linesBlocksService.getEditableTextLineFromNode(
@@ -39,7 +46,7 @@ class DocumentService {
 
         // Block
       } else if (node is BlockM) {
-        docElems.add(
+        docElements.add(
           Directionality(
             textDirection: getDirectionOfNode(node),
             child: _editableTextBlock(
@@ -57,8 +64,23 @@ class DocumentService {
       }
     }
 
-    return docElems;
+    return docElements;
   }
+
+  // Right before sending the document to the rendering layer, we want to check if the document is not empty.
+  // If the document is empty, we replace it with placeholder content.
+  DocumentM getDocOrPlaceholder(EditorState state) =>
+      state.document.document.isEmpty() &&
+              state.editorConfig.config.placeholder != null
+          ? DocumentM.fromJson(
+              jsonDecode(
+                '[{'
+                '"attributes":{"placeholder":true},'
+                '"insert":"${state.editorConfig.config.placeholder}\\n"'
+                '}]',
+              ),
+            )
+          : state.document.document;
 
   Widget _editableTextBlock(
     BlockM node,